Add `move` to Robot class
Add move function to the Robot class, which will move the Robot forward
one space in the direction it is facing and return an updated Robot
instance with said change. This works for its equivalent `exec` command
as well.

Also, the Robot cannot fall off the table edges at any
coordinate/position less than 0 or greater than 4, as per the Toy Robot
requirements.